requirements.txt
python-dotenv==1.0.1
validators==0.34.0
peewee==3.17.6
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
